Converted config file style to look more like normal Ruby
Previously you used a block (with nothing yielded into it) to define
style.  While this looked "pretty", it felt like a diversion of normal
Ruby, so I changed it to yield an OpenStruct.  Now, using the config
file feels more like Ruby and when we get to allowing users to define
their own Rulers, the config file will be ready to take options for it.